requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
Django==3.2.15
idna==3.3
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
typing_extensions==4.3.0
urllib3==1.26.11
psycopg2>=2.8
appdirs==1.4.4
argcomplete==2.0.0
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
Django==3.2.15
future==0.18.2
idna==3.3
importlib-metadata==4.12.0
Pillow==9.2.0
pycups==2.0.1
pyserial==3.5
python-barcode==0.14.0
python-printer-escpos==0.0.4
pytz==2022.1
pyusb==1.2.1
PyYAML==6.0
qrcode==7.3.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.3.0
urllib3==1.26.11
viivakoodi==0.8.0
zipp==3.8.1